refactor: contracts in repository
