+++ src/domain/application/use-cases/pizza/delete-pizza.ts
@@ -1,4 +1,4 @@
-import { PizzaRepository } from "../../repositories/pizza-repository";
+import { PizzaRepository } from "../../repositories/product-repository";
 import { Either, left, right } from "../../../../core/either";
 import { ResourceNotFoundError } from "../../../../core/errors/resource-not-found-error";
