+++ src/domain/enterprise/entities/index.ts
@@ -2,4 +2,4 @@ export * from './address'
 export * from './customer'
 export * from './order'
 export * from './product'
-export * from './drink'
+export * from './productOrder'
