+++ src/infra/repository/prisma/prisma-pizza.ts
@@ -1,5 +1,5 @@
 import { prisma } from "../../../lib/prisma";
-import { PizzaRepository } from "../../../domain/application/repositories/pizza-repository";
+import { PizzaRepository } from "../../../domain/application/repositories/product-repository";
 import { Pizza } from "../../../domain/enterprise/entities";
 
 export class PrismaPizzaRepository implements PizzaRepository {
